refactor(profiler/kineto): share the arch-detection util between code and tests
