+++ test/utils.py
@@ -1,4 +1,3 @@
-import functools
 import multiprocessing
 import os
 import random
@@ -8,6 +7,7 @@ import numpy as np
 import pytest
 import torch
 
+from torch_rbln._internal.device_arch_utils import is_atom_device, is_rebel_device
 from torch_rbln._internal.ops_utils import SupportedDtypes
 
 
@@ -87,42 +87,6 @@ def requires_physical_devices(num_devices):
         physical_device_count < num_devices,
         reason=f"Requires at least {num_devices} physical devices, found {physical_device_count}",
     )
-
-
-def _arch_from_npu_name(name: str) -> str:
-    """Map an RBLN NPU name to a device family: ``RBLN-CA*`` -> ``"atom"``,
-    ``RBLN-CR*`` -> ``"rebel"``, anything else -> ``"unknown"``.
-    """
-    name = name.upper()
-    if name.startswith("RBLN-CA"):
-        return "atom"
-    if name.startswith("RBLN-CR"):
-        return "rebel"
-    return "unknown"
-
-
-@functools.lru_cache(maxsize=1)
-def get_device_arch() -> str:
-    """Identify the current NPU family (``"atom"``/``"rebel"``/``"unknown"``) via
-    ``get_npu_name`` from ``rebel-compiler`` (cached). Returns ``"unknown"`` if the
-    NPU name can't be queried.
-    """
-    try:
-        from rebel.device_info import get_npu_name
-
-        return _arch_from_npu_name(get_npu_name(0) or "")
-    except Exception:
-        return "unknown"
-
-
-def is_atom_device() -> bool:
-    """True on the ATOM lineup (``RBLN-CA*``); thin wrapper over :func:`get_device_arch`."""
-    return get_device_arch() == "atom"
-
-
-def is_rebel_device() -> bool:
-    """True on the REBEL lineup (``RBLN-CR*``); thin wrapper over :func:`get_device_arch`."""
-    return get_device_arch() == "rebel"
 
 
 def xfail_atom(reason: str):
