+++ torch_rbln/__init__.py
@@ -194,14 +194,9 @@ def _create_process_group_rbln(dist_backend_opts, pg_options):
 
 def _initialize_kineto_profiler() -> None:
     """Register the rbln torch.profiler (kineto) bridge."""
-    try:
-        from rebel.device_info import get_npu_name
+    from torch_rbln._internal.device_arch_utils import is_atom_device
 
-        npu = (get_npu_name(0) or "").upper()
-    except Exception:
-        npu = ""
-
-    if npu.startswith("RBLN-CA"):  # ATOM
+    if is_atom_device():
         warnings.warn("rbln torch.profiler activities are not supported on ATOM and will be skipped", stacklevel=2)
         return
